pyproject.toml
[tool.poetry]
name = "mallard"
version = "0.2"
description = "MAchine Learning Large ARtifact Database"
authors = ["Daniel Petti <djpetti@gmail.com>", "Justin Hong <wjtw9802@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "~3.9"
minio = "^7.0.1"
mysql-connector = "^2.2.9"
pydantic = "^1.7.3"
fastapi = "^0.63.0"
uvicorn = {extras = ["standard"], version = "^0.13.3"}
python-multipart = "^0.0.5"
Pillow = "^8.1.0"
confuse = "^1.4.0"
loguru = "^0.5.3"
exifread = "^2.3.2"
aioitertools = "^0.7.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.3"
pre-commit = "^2.6.0"
pytest = "^6.0.1"
pytest_mock = "^3.3.1"
pytest_cov = "^2.10.1"
pytest_black = "^0.3.11"
pytest_flake8 = "^1.0.6"
faker = "^5.6.0"
pytest-asyncio = "^0.14.0"
pytest_faker = "^2.0.0"
python-dateutil = "^2.8.1"
requests = "^2.25.1"

[tool.black]
line-length = 79
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.venv
    | build
    | dist
  )/
)
'''

[tool.isort]
profile = "black"
line_length = 79
skip = "('.eggs', '.git', '.venv', 'build', 'dist')"

[tool.kedro]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
